+++ apps/ceo-front/src/app/login/components/RegisterForm.tsx
@@ -1,0 +1,5 @@
+const RegisterForm = () => {
+  return <></>;
+};
+
+export default RegisterForm;
