+++ apps/ceo-front/src/app/login/page.tsx
@@ -1,190 +1,18 @@
 'use client';
 import Image from 'next/image';
-import { useRouter, useSearchParams } from 'next/navigation';
-import { JSX, useCallback, useState, useEffect } from 'react';
+import { useState } from 'react';
 
-import { AuthSocialLoginBody, SocialPlatform, socialPlatforms } from '@repo/api/auth';
-import { STORAGE_KEYS } from '@repo/api/configs/storage-keys';
-import { SocialButton } from '@repo/design-system/components';
-import { Input, Checkbox, Button } from '@repo/design-system/components/b2c';
-import {
-  AppleIcon,
-  GoogleIcon,
-  KakaoIcon,
-  NaverIcon,
-  EyeOffIcon,
-  EyeOnIcon,
-} from '@repo/design-system/icons';
-import Cookies from 'js-cookie';
-
+import LoginForm from './components/LoginForm';
+import RegisterForm from './components/RegisterForm';
 import FooterLayout from '@/components/Layouts/FooterLayout';
 import { useAuth } from '@/components/Provider/AuthProvider';
-import { env } from '@/configs/env';
-
-// TODO 컴포넌트 분리
-const platformIcons: Record<SocialPlatform, JSX.Element> = {
-  KAKAO: <KakaoIcon size={30} />,
-  NAVER: <NaverIcon size={30} />,
-  APPLE: <AppleIcon size={30} fill='white' />,
-  GOOGLE: <GoogleIcon size={30} />,
-};
-
-const platformStyles: Record<SocialPlatform, string> = {
-  KAKAO: 'bg-[#FEE500] rounded-[50%] h-[64px] w-[64px] ',
-  NAVER: 'bg-[#03C75A] rounded-[50%] h-[64px] w-[64px]',
-  APPLE: 'bg-black rounded-[50%] h-[64px] w-[64px]',
-  GOOGLE: 'bg-[#F2F2F2] rounded-[50%] h-[64px] w-[64px]',
-};
-
-const getPlatformUrl = (platform: SocialPlatform): string | null => {
-  const {
-    GOOGLE_CLIENT_ID,
-    OAUTH_REDIRECT_URI,
-    KAKAO_CLIENT_ID,
-    NAVER_CLIENT_ID,
-    APPLE_CLIENT_ID,
-  } = env;
-
-  const platformUrls: Record<SocialPlatform, string> = {
-    GOOGLE: `https://accounts.google.com/o/oauth2/v2/auth?client_id=${GOOGLE_CLIENT_ID}&redirect_uri=${OAUTH_REDIRECT_URI}&response_type=code&scope=email%20profile`,
-    KAKAO: `https://kauth.kakao.com/oauth/authorize?response_type=code&client_id=${KAKAO_CLIENT_ID}&redirect_uri=${OAUTH_REDIRECT_URI}`,
-    NAVER: `https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=${NAVER_CLIENT_ID}&redirect_uri=${OAUTH_REDIRECT_URI}&state=test1234`,
-    APPLE: `https://appleid.apple.com/auth/authorize?response_type=code&response_mode=form_post&client_id=${APPLE_CLIENT_ID}&redirect_uri=${OAUTH_REDIRECT_URI}&scope=name%20email&state=test123`,
-  };
-
-  return platformUrls[platform];
-};
-
-const openSocialLoginPopup = (loginUrl: string) => {
-  const width = 500;
-  const height = 600;
-  const left = window.screenX + (window.outerWidth - width) / 2;
-  const top = window.screenY + (window.outerHeight - height) / 2;
-
-  return window.open(
-    loginUrl,
-    'login-popup',
-    `width=${width},height=${height},left=${left},top=${top}`,
-  );
-};
+import { useLoginForm } from '@/hooks/useLoginForm';
 
 export default function LoginPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const [, setIsLoading] = useState<boolean>(false);
-  const [email, setEmail] = useState<string>('');
-  const [password, setPassword] = useState<string>('');
-  const [showPassword, setShowPassword] = useState<boolean>(false);
-  const [errorMessage, setErrorMessage] = useState<string>('');
-  const [isRememberId, setIsRememberId] = useState(false);
+  const [isSignUp, setIsSignUp] = useState(false);
 
-  const { login, socialLogin } = useAuth();
-
-  const handleTogglePasswordVisibility = () => {
-    setShowPassword((prev) => !prev);
-  };
-
-  const handleLogin = async () => {
-    if (!email || !password) {
-      setErrorMessage('이메일과 비밀번호를 입력해주세요');
-      return;
-    }
-    setIsLoading(true);
-    try {
-      await login({ email, password, role: 'CEO' });
-
-      if (isRememberId) {
-        localStorage.setItem('saved_email', email);
-      } else {
-        localStorage.removeItem('saved_email');
-      }
-
-      router.replace('/');
-    } catch (err: any) {
-      // TODO any 타입 fix
-      const status = err?.response?.status;
-      const message = err?.response?.data?.message;
-
-      if (status === 400) {
-        setErrorMessage('비밀번호가 일치하지 않습니다.\n입력한 내용을 다시 확인해주세요');
-      } else if (status === 401 || status === 404) {
-        setErrorMessage('존재하지 않는 계정입니다.\n이메일을 다시 확인해주세요');
-      } else if (status === 500) {
-        setErrorMessage('서버 오류가 발생했습니다.\n잠시 후 다시 시도해주세요');
-      } else {
-        setErrorMessage(message ?? '로그인에 실패했습니다.\n다시 시도해주세요');
-      }
-    }
-    setIsLoading(false);
-  };
-
-  const handleSocialLogin = useCallback(
-    async (platform: SocialPlatform) => {
-      const loginUrl = getPlatformUrl(platform);
-      if (!loginUrl) return;
-
-      const popup = openSocialLoginPopup(loginUrl);
-      if (!popup) return;
-
-      const handleMessage = async (event: MessageEvent) => {
-        const allowedOrigins = [
-          'http://localhost:3000',
-          'https://ceo-stage.fooding.im',
-          'https://ceo.fooding.im',
-        ];
-
-        if (!allowedOrigins.includes(event.origin)) {
-          console.warn(`Origin mismatch: ${event.origin} !== any of`, allowedOrigins);
-          return;
-        }
-
-        if (event.data?.action === 'closePopup') {
-          popup.close?.();
-          return;
-        }
-
-        const { code } = event.data;
-        if (!code) return;
-        window.removeEventListener('message', handleMessage);
-
-        setIsLoading(true);
-        try {
-          const credentials: AuthSocialLoginBody = {
-            code,
-            provider: platform,
-            redirectUri: env.OAUTH_REDIRECT_URI || '',
-            role: 'CEO',
-          };
-          await socialLogin(credentials);
-
-          Cookies.set(STORAGE_KEYS.RECENT_PROVIDER, platform, {
-            path: '/',
-            secure: process.env.NODE_ENV === 'production',
-            sameSite: 'lax',
-          });
-
-          const returnTo = searchParams.get('returnTo') || '/';
-          router.replace(returnTo);
-        } catch (error) {
-          console.error('Social login failed:', error);
-        } finally {
-          setIsLoading(false);
-          window.removeEventListener('message', handleMessage);
-        }
-      };
-
-      window.addEventListener('message', handleMessage);
-    },
-    [router, searchParams, socialLogin],
-  );
-
-  useEffect(() => {
-    const savedEmail = localStorage.getItem('saved_email');
-    if (savedEmail) {
-      setEmail(savedEmail);
-      setIsRememberId(true);
-    }
-  }, []);
+  const { login } = useAuth();
+  const loginForm = useLoginForm(login);
 
   return (
     <div className='flex flex-col'>
@@ -220,77 +48,7 @@ export default function LoginPage() {
               <p className='body-6 text-blue-500'>계정이 없으신가요?</p>
             </div>
             <div className='h-[45px]' />
-            <form>
-              <div className='flex flex-col gap-[20px]'>
-                <div className='flex flex-col gap-[8px]'>
-                  <p className='subtitle-3'>아이디</p>
-                  <Input
-                    required
-                    autoFocus
-                    type='email'
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    className={`px-[20px] py-[18px] body-2 h-[58px] ${errorMessage ? 'border-red-500 text-red-500' : ''}`}
-                  />
-                </div>
-                <div className='flex flex-col gap-[8px]'>
-                  <p className='subtitle-3'>비밀번호</p>
-                  <div className='relative'>
-                    <Input
-                      required
-                      type={showPassword ? 'text' : 'password'}
-                      value={password}
-                      onChange={(e) => setPassword(e.target.value)}
-                      className={`px-[20px] py-[18px] body-2 h-[58px] ${errorMessage ? 'border-red-500 text-red-500' : ''}`}
-                    />
-                    <button
-                      type='button'
-                      onClick={handleTogglePasswordVisibility}
-                      className='absolute right-[20px] top-1/2 -translate-y-1/2 text-gray-500 cursor-pointer hover:text-gray-700'
-                      aria-label={showPassword ? '비밀번호 숨기기' : '비밀번호 표시'}
-                    >
-                      {showPassword ? <EyeOnIcon /> : <EyeOffIcon />}
-                    </button>
-                  </div>
-                </div>
-              </div>
-              <div className='h-[12px]' />
-              <div className='flex justify-between'>
-                <div className='flex gap-[8px] items-center cursor-pointer'>
-                  <Checkbox
-                    checked={isRememberId}
-                    onChange={(checked) => setIsRememberId(Boolean(checked))}
-                  />
-                  <p onClick={() => setIsRememberId((prev) => !prev)}>아이디 저장</p>
-                </div>
-
-                <p className='body-6 text-gray-5'>아이디/비밀번호를 잊으셨나요?</p>
-              </div>
-              {errorMessage.length > 0 && (
-                <p className='mt-2 text-sm text-red-500 whitespace-pre-line'>{errorMessage}</p>
-              )}
-              <div className='h-[68px]' />
-              <Button
-                onClick={() => handleLogin()}
-                disabled={email.length == 0 || password.length == 0}
-                className='py-[17px] rounded-full subtitle-1'
-              >
-                로그인
-              </Button>
-            </form>
-            <div className='h-[120px]' />
-            {/* 소셜 로그인 버튼 */}
-            <div className='flex justify-center gap-4 mt-8'>
-              {socialPlatforms.map((platform) => (
-                <SocialButton
-                  key={platform}
-                  platform={platform}
-                  icon={platformIcons[platform]}
-                  onClick={() => handleSocialLogin(platform)}
-                  styles={platformStyles[platform]}
-                />
-              ))}
-            </div>
+            {isSignUp ? <RegisterForm /> : <LoginForm {...loginForm} />}
           </div>
         </div>
